Assistant checkpoint: Create e-book library system with metadata management
Assistant generated file changes:
- data/books.json: Create books metadata storage
- data/lists.json: Create lists for authors, genres, series
- aniol.php: Create book manager functionality
- ksiazki.html: Create initial library page

---

User prompt:

Great. I would like to create a simple e-book library. Assumptions: all files will be stored in _ksiazki folder; There will be aniol.php that is the bookmanager and ksiazki.html which will be the output (and will be static file generated by aniol.php). All linkes should be domain indepedent and relative.  I will also like to have metadata for each book stored in a json format (it can be one file for all books), for each book we will have the following data: file_name (not editable), date_uploaded (not editable), title, author (or authors) with first name, last name, genre (may be multiple), series (if aplicable), position in series (if aplicable). I would like to be able to choose the authors, genre, series from a list of already added positions (they should be also kept in some handy json file, and not be generated on the fly everysingle time), or be able to add new author / genre / series if needed