Assistant checkpoint
Assistant generated file changes:
- aniol.php: Add tabs, pagination, and HTML generation button

---

User prompt:

The main goal of the aniol.php is to be able to add data on books that have been newly uploaded to the folder, and to generate the ksiazki.html file (on demand). In the status bar add information when the ksiazki.html has been updated last. Add also a button that will run a generation of the ksiazki.html file; in aniol.php create a tab div. One tab should be called "Nieopisane" in which we will show books without metadata, and "gotowe" with all books with metadata. Both tabs should show lists of 10 positions, with pagination if necessary